requirements.txt
streamlit
pymongo
python-dotenv
plotly
pandas